Move event system codes to a separated file event.py
