+++ plasma_cash/child_chain/__init__.py
@@ -1,19 +1,5 @@
 from flask import Flask
 
-callback = {}
-
-
-def on(event):
-    def decorator(func):
-        callback[event] = func
-        return func
-    return decorator
-
-
-def emit(event, *arg, **kw):
-    if event in callback and callable(callback[event]):
-        callback[event](*arg, **kw)
-
 
 def create_app(is_unit_test=False):
     app = Flask(__name__)
